add validDHSmallFlag and checkSmallDHKeys
