+++ engines/pixiedust/parser.go
@@ -55,19 +55,19 @@ func FlagSettings() []argparser.Flag {
 	return []argparser.Flag{
 		{ID: 0, Desc: "Pixie Dust\nE.g., $ sudo ./offscan pixie <FLAGS>"},	
 		{ID: jobs,    Short: "j", Long: "jobs",    HasValue: true, Desc: "Number of workers"},
-		{ID: pke ,    Short: "e", Long: "pke",     HasValue: true, Req: true, Desc: "Public Key Enrollee"},
-		{ID: pkr ,    Short: "r", Long: "pkr",     HasValue: true, Req: true, Desc: "Public Key Registrar"},
-		{ID: eHash1,  Short: "1", Long: "ehash1",  HasValue: true, Req: true, Desc: "Enrollee Hash 1"},
-		{ID: eHash2,  Short: "2", Long: "ehash2",  HasValue: true, Req: true, Desc: "Enrollee Hash 2"},
-		{ID: authKey, Short: "a", Long: "authkey", HasValue: true, Req: true, Desc: "Authentication Session Key"},
-		{ID: eNonce,  Short: "n", Long: "enonce",  HasValue: true, Req: true, Desc: "Enrollee Nonce"},
+		{ID: pke ,    Short: "e", Long: "pke",     HasValue: true, Desc: "Public Key Enrollee"},
+		{ID: pkr ,    Short: "r", Long: "pkr",     HasValue: true, Desc: "Public Key Registrar"},
+		{ID: eHash1,  Short: "1", Long: "ehash1",  HasValue: true, Desc: "Enrollee Hash 1"},
+		{ID: eHash2,  Short: "2", Long: "ehash2",  HasValue: true, Desc: "Enrollee Hash 2"},
+		{ID: authKey, Short: "a", Long: "authkey", HasValue: true, Desc: "Authentication Session Key"},
+		{ID: eNonce,  Short: "n", Long: "enonce",  HasValue: true, Desc: "Enrollee Nonce"},
 		{ID: rNonce,  Short: "m", Long: "rnonce",  HasValue: true, Desc: "Registrar Nonce"},
 		{ID: ebssid,  Short: "b", Long: "ebssid",  HasValue: true, Desc: "Enrollee MAC"},
 		{ID: m5enc,   Short: "5", Long: "m5enc",   HasValue: true, Desc: "DH Small"},
 		{ID: m7enc,   Short: "7", Long: "m7enc",   HasValue: true, Desc: "DH Small"},
 		{ID: force,   Short: "f", Long: "force",   Desc: "Force bruteforce"},
 		{ID: dhSmall, Short: "S", Long: "dhsmall", Desc: "Use small DH group"},
-		{ID: modes,    Long: "mode",   HasValue: true, Desc: "Attack mode (1-5, 0=auto)"},
+		{ID: modes,   Long: "mode",   HasValue: true, Desc: "Attack mode (1-5, 0=auto)"},
 		{ID: start,   Long: "start",  HasValue: true, Desc: "Start timestamp for mode 3"},
 		{ID: end,     Long: "end",    HasValue: true, Desc: "End timestamp for mode 3"},
 		{ID: cStart,  Long: "cstart", HasValue: true, Desc: "Custom start seed"},
